# Use System.Text.Json in ColorPickerValueConverter