### Prerequisites

- [x] I have added steps to test this contribution in the description below

### Description
Improving performance of the ColorPickerValueConverter by using System.Text.Json instead of Newtonsoft.Json

### How to test
- On a content type create a property using the ColorPicker editor
- Call this property in a template and make sure the values are getting displayed correctly

### Benchmark
```
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.2130 (21H2)
Intel Core i9-10885H CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-rc.2.22477.23
  [Host]     : .NET 6.0.10 (6.0.1022.47605), X64 RyuJIT
  DefaultJob : .NET 6.0.10 (6.0.1022.47605), X64 RyuJIT


| Method |     Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------- |---------:|----------:|----------:|-------:|----------:|
|    New | 2.806 us | 0.0561 us | 0.0468 us | 0.0229 |     200 B |
|    Old | 4.856 us | 0.0960 us | 0.1682 us | 0.5341 |   4,496 B |
```