@nul800sebastiaan 2023-03-07
Hi @CyberReiter - Sorry for the late response, we're working on our backlog at the moment.

I will close this one and the related Newtonsoft updates. There are subtle differences between Newtonsoft and System.Text.Json that will lead to unexpected output and breaking things.

We are planning to remover Newtonsoft completely from the solution in Umbraco 13: https://github.com/umbraco/Announcements/issues/9

Thanks for the PR!